Show rrt

+++ network/src/main/java/io/bitsquare/p2p/network/Statistic.java
@@ -3,9 +3,7 @@ package io.bitsquare.p2p.network;
 import io.bitsquare.common.UserThread;
 import io.bitsquare.p2p.Message;
 import javafx.beans.property.IntegerProperty;
-import javafx.beans.property.LongProperty;
 import javafx.beans.property.SimpleIntegerProperty;
-import javafx.beans.property.SimpleLongProperty;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -51,7 +49,7 @@ public class Statistic {
     private final IntegerProperty receivedBytes = new SimpleIntegerProperty(0);
     private final Map<String, Integer> receivedMessages = new ConcurrentHashMap<>();
     private final Map<String, Integer> sentMessages = new ConcurrentHashMap<>();
-    private final LongProperty roundTripTime = new SimpleLongProperty(0);
+    private final IntegerProperty roundTripTime = new SimpleIntegerProperty(0);
 
 
     ///////////////////////////////////////////////////////////////////////////////////////////
@@ -103,7 +101,7 @@ public class Statistic {
         sentMessages.put(messageClassName, counter);
     }
 
-    public void setRoundTripTime(long roundTripTime) {
+    public void setRoundTripTime(int roundTripTime) {
         this.roundTripTime.set(roundTripTime);
     }
 
@@ -139,7 +137,7 @@ public class Statistic {
         return creationDate;
     }
 
-    public LongProperty roundTripTimeProperty() {
+    public IntegerProperty roundTripTimeProperty() {
         return roundTripTime;
     }
